refactored Play and Team VC
